refactor(layout): centralize paint ordering
Split the remaining large CSS and inline-layout implementations into cohesive same-package files, then make PDF and raster adapters consume one layout-owned paint-order policy. Preserve existing operation ordering, package-private seams, and backend-specific drawing behavior.
